Updated stb_image.h to v2.06, removed more gl.h dependencies, fixed UI textures

+++ WargameEngine/WargameEngine/WargameEngine/UI/UIList.h
@@ -24,7 +24,7 @@ public:
 	virtual void SetTheme(std::shared_ptr<CUITheme> theme) override;
 private:
 	std::vector<std::wstring> m_items;
-	int m_selected;
+	size_t m_selected;
 	std::function<void()> m_onChange;
 	CUIScrollBar m_scrollbar;
 };
